Create replace.clade.R
function to replace a clade in a tree. This if for adding taxa via pastis to very large trees that are too big for pastis to handle. The clade is extracted, run through pastis and MrBayes, then this function finds the old clade in the original tree, and replaces it with the new one. This works for one tree, or a distribution of pseudoposterior trees